+++ Editor/AssetExtractors/ApdAssetPostprocessor.cs
@@ -1,28 +1,27 @@
-using System.IO;
+using System.Diagnostics.CodeAnalysis;
+using System.Linq;
 using UnityEditor;
-using AppodealStack.UnityEditor.InternalResources;
+using AppodealStack.UnityEditor.Utils;
 
-// ReSharper Disable CheckNamespace
+// ReSharper Disable once CheckNamespace
 namespace AppodealStack.UnityEditor.AssetExtractors
 {
     internal class ApdAssetPostprocessor : AssetPostprocessor
     {
-#if UNITY_2021_3_OR_NEWER
+        [SuppressMessage("ReSharper", "UnusedMember.Local")]
+        [SuppressMessage("ReSharper", "UnusedParameter.Local")]
+        [SuppressMessage("ReSharper", "Unity.IncorrectMethodSignature")]
         private static void OnPostprocessAllAssets(string[] importedAssets, string[] deletedAssets, string[] movedAssets, string[] movedFromAssetPaths, bool didDomainReload)
         {
-#else
-        private static void OnPostprocessAllAssets(string[] importedAssets, string[] deletedAssets, string[] movedAssets, string[] movedFromAssetPaths)
-        {
-            const string prefsPath = "Assets/Appodeal/Editor/InternalResources/PluginPreferences.asset";
-            if (File.Exists(prefsPath) && AssetDatabase.LoadAssetAtPath<PluginPreferences>(prefsPath) == null) return;
+#if APPODEAL_DEV
+            return;
 #endif
+            if (deletedAssets.Any(asset => asset.Contains(AppodealEditorConstants.PackageDir))) return;
+
             if (AndroidLibraryInstaller.InstallAndroidLibrary() | AppodealAdaptersInstaller.InstallAdapters())
             {
                 AssetDatabase.Refresh();
             }
-
-            AssetDatabase.SaveAssetIfDirty(PluginPreferences.Instance);
-
         }
     }
 }