+++ Editor/DependencyManager/Constants/DmConstants.Uss.cs
@@ -1,0 +1,12 @@
+// ReSharper disable once CheckNamespace
+namespace AppodealInc.Mediation.DependencyManager.Editor
+{
+    public static partial class DmConstants
+    {
+        internal static class Uss
+        {
+            public const string SdkClassName = "sdk";
+            public const string MediationIconClassName = "ad-network-sdk-card__mediation-icon";
+        }
+    }
+}